fix(allocations): Assigned regions only to selected users
