+++ app/locations/middleware.js
@@ -1,7 +1,5 @@
 const { find, get } = require('lodash')
 
-const referenceDataService = require('../../common/services/reference-data')
-
 function setUserLocations(req, res, next) {
   req.userLocations = get(req.session, 'user.locations', [])
   next()
@@ -49,20 +47,10 @@ function setAllLocations(req, res, next) {
   next()
 }
 
-async function setRegions(req, res, next) {
-  try {
-    req.session.regions = await referenceDataService.getRegions()
-    next()
-  } catch (error) {
-    next(new Error('Failed to retrieve the regions'))
-  }
-}
-
 module.exports = {
   setUserLocations,
   checkLocationsLength,
   setLocation,
   setRegion,
   setAllLocations,
-  setRegions,
 }
